oeqa/sdk/cases/autotools.py: fix host_sys value to handle multilib case
For now, testsdk always fails for multilib. This could be reproduced
by the following steps:

  1. Add to local.conf:
     OE_FRAGMENTS += "machine/qemux86-64 distro/poky"
     OE_FRAGMENTS += "core/yocto-autobuilder/multilib-x86-lib32"
     IMAGE_CLASSES += "testsdk"
  2. bitbake core-image-minimal -c populate_sdk && \
     bitbake core-image-minimal -c testsdk

This is because in case of multilib, HOST_SYS value is not correct.

So switch to use CONFIGURE_FLAGS environment variable to extract
the host_sys value.

(From OE-Core rev: b0e42adf3f122eb1ad46afffaffd180613148701)

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>
Signed-off-by: Mathieu Dubois-Briand <mathieu.dubois-briand@bootlin.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>